Refactor Range serialization with RangeModel adapter
Introduced a new RangeModel class to encapsulate serialization logic for Range and RedisZSetCommands.Range. Updated RangeSerializer and RedisZSetCommandsRangeSerializer to use RangeModel, improving code reuse and maintainability.
